document_parser/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-pptx
openpyxl
pytesseract
Pillow
openai
python-dotenv